+++ BARICA-u-oblacima/backendapi.py
@@ -6,6 +6,10 @@ import os
 import sys
 import subprocess
 import pty
+
+
+import socket, errno
+import threading
 from time import sleep
 
 DEBUG = False
@@ -13,16 +17,82 @@ DEBUG = False
 # Used ports for streams
 STREAM_PORTS = []
 
-def get_stream_port():
-    ''' TODO: make ports rotate, now it will
-        go from 2000 until forever. '''
+# Port manager thread stop flag
+PORT_MANAGER_STOP = 0
+
+#semaphore
+threadLock = threading.Lock()
+
+# STREAM_PORT manager in particular thread
+class StreamPortManager(threading.Thread):
+    def __init__(self, ip, timeDelay):
+        threading.Thread.__init__(self)
+        self.ip = ip
+        self.timeDelay = timeDelay #number of cycles between port processing
+        self.timeCount = timeDelay #cycle counter
+        print("Port manager started")
+
+    def run(self):
+        global PORT_MANAGER_STOP
+        while 1:
+            if PORT_MANAGER_STOP == 1:
+                break
+            if(self.timeCount == self.timeDelay):
+                self.portProcessing()
+                self.timeCount = 0
+            else:
+                self.timeCount +=1
+            sleep(0.5) # one cycle is 500 ms (0.5 s)
+        
+    def stop(self):
+        global PORT_MANAGER_STOP
+        PORT_MANAGER_STOP = 1
+        print("Port manager stopping. Please wait for thread to exit.")
+        
+
+    def portProcessing(self):
+        global STREAM_PORTS
+        threadLock.acquire()
+        print("Port processing...\n used ports: {0:d}".format(len(STREAM_PORTS)))
+        for i in range(0,len(STREAM_PORTS)):
+            if i < 0 or i >= len(STREAM_PORTS):
+                threadLock.release()
+                return
+            p = STREAM_PORTS[i]
+            if check_port(self.ip, p) == 1:
+                print("IP: {0}, port {1:d} {2:s}".format(self.ip, p, "CLOSED"))
+                del STREAM_PORTS[i]
+                i -= 1
+            else:
+                print("IP: {0}, port {1:d} {2:s}".format(self.ip, p, "OPEN"))
+        threadLock.release()
+
+
+def get_stream_port(ip):
     global STREAM_PORTS
-    if STREAM_PORTS:
-        port = STREAM_PORTS[ -1 ] + 1
-    else:
-        port = 2000
-    STREAM_PORTS.append( port )
-    return port
+    result = -1
+    threadLock.acquire()
+    for i in range(2000,3000):
+        if i in STREAM_PORTS:
+            continue
+        if check_port(ip,i) == 1:
+            STREAM_PORTS.append(i)
+            result = i
+            break
+    threadLock.release()
+    print("Free port found: {0:d}".format(result))
+    return result
+
+def check_port(ip, port):
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    try:
+        s.bind((ip,port))
+    except socket.error as e:
+        return 0
+    finally:        
+        s.close()
+    return 1   
+
 
 
 # Stolen from https://stackoverflow.com/questions/229186/os-walk-without-digging-into-directories-below
@@ -43,10 +113,11 @@ class ModuleException( Exception ):
             
 # Module class
 class Module:
-    def __init__( self, name, descriptor, location ):
+    def __init__( self, name, descriptor, location, ip):
         self.name = name
         self.descriptor = descriptor
         self.location = location
+        self.ip = ip
 
     def __str__( self ):
         return json.dumps( self.descriptor, indent=4, sort_keys=True )
@@ -62,7 +133,8 @@ class Module:
         '''
         if self.descriptor[ "methods" ][ method ][ "stdin_stream" ]:
             if self.descriptor[ "methods" ][ method ][ "stdout_stream" ]:
-                port = get_stream_port()
+                #port = get_stream_port()
+                port = get_stream_port(self.ip)
                 command = [ "nc", "-lp", str( port ), "|", os.path.join( self.location, self.descriptor[ "methods" ][ method ][ "run" ] ), method, "|", "nc", args[ 0 ], str( args[ 1 ] ) ]
                 try:
                     proc = subprocess.Popen( ' '.join( command ), shell=True )
@@ -80,7 +152,7 @@ class Module:
                 except Exception as e:
                     raise ModuleException( "Error while running module %s. Error was:\n%s" % ( self.name, str( e ) ) )
             
-def load_modules():
+def load_modules(ip):
     bar = tqdm( walklevel( 'modules', level=1 ) )
     modules = {}
     
@@ -95,7 +167,7 @@ def load_modules():
             if fname == 'interface.json':
                 mod = json.loads( open( full_path ).read() )
                 printt( "Processing %s" % mod[ 'name' ] )
-                newmod = Module( mod[ "name" ], mod, dir_path )
+                newmod = Module( mod[ "name" ], mod, dir_path, ip)
                 modules[ mod[ 'name' ] ] = newmod
                 for m in mod[ "methods" ]:
                     method = mod[ "methods" ][ m ]
@@ -125,4 +197,4 @@ def load_modules():
 
 
 if __name__ == '__main__':
-    load_modules()
+    load_modules('localhost')
